requirements.txt
Flask==2.1.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-SocketIO==5.1.1
Flask-Login==0.6.0
Flask-Limiter==2.7.0
python-socketio==5.4.0
eventlet==0.33.0
psycopg2-binary==2.9.3
SQLAlchemy==1.4.32
Werkzeug==2.0.3
redis==4.5.1
python-dotenv==0.21.0
